refactor: moved message and modal package to builder package, and separate message file
builder로 사용되는 패키지를 묶기 위해 builder 패키지를 만들어 옮겼고, 메시지 파일의 크기가 비대해져서 나눴습니다.
builder/ 내부 이외의 파일들은 변경에 따른 변화입니다.
